feature/ASE-109: Added Spring Data JPA Audit

+++ src/main/java/com/bloxico/ase/userservice/dto/entity/evaluation/CountryEvaluationDetailsDto.java
@@ -9,10 +9,10 @@ import lombok.EqualsAndHashCode;
 public class CountryEvaluationDetailsDto {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("country_id")
-    private Integer countryId;
+    private Long countryId;
 
     @JsonProperty("price_per_evaluation")
     private Integer pricePerEvaluation;
