+++ src/main/java/com/bloxico/ase/userservice/repository/address/CountryRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
 
     Optional<Country> findByNameIgnoreCase(String name);
 
-    int countByRegionsIdEquals(int regionId);
+    int countByRegionsIdEquals(Long regionId);
 
 }
