+++ src/test/java/com/bloxico/ase/testutil/UtilLocation.java
@@ -67,7 +67,7 @@ public class UtilLocation {
         return MAPPER.toDto(savedCountry());
     }
 
-    public CountryDto genCountryDto(int id, String name, Set<RegionDto> regions) {
+    public CountryDto genCountryDto(Long id, String name, Set<RegionDto> regions) {
         var dto = new CountryDto();
         dto.setId(id);
         dto.setName(name);
